move ledger explorer logic in ledger file
